Fix schedule-task delete

+++ src/command/schedule_task/delete/executer.rs
@@ -39,9 +39,14 @@ impl Executer {
     pub fn run(&self, rule_name: &str) -> Result<(), Box<error::Error>> {
         trace!("command::schedule_task::delete::Executer::run");
 
-        try!(self.delete_rule(rule_name));
+        if try!(self.rule_exists(rule_name)) {
+            try!(self.remove_targets(rule_name));
+            try!(self.delete_rule(rule_name));
+            output::PrintLine::success("Finished deleting the scheduled task");
+        } else {
+            output::PrintLine::success("The rule does not exists");
+        }
 
-        output::PrintLine::success("Finished deleting the scheduled task");
         Ok(())
     }
 }
